requirements.txt
numpy>=1.24.0
pyvista>=0.40.0
pyvistaqt>=0.9.0
PyQt6>=6.4.0
roboticstoolbox-python>=1.0.0

